+++ core/src/main/scala/mongo4s/internal/DirectMongoCollectionImpl.scala
@@ -21,7 +21,7 @@ import mongo4s.changestream.{ChangeEvent, WatchOptions}
 import mongo4s.queries.{AggregateQuery, DistinctQuery, FindQuery}
 import mongo4s.{Effect, Field, MongoCollection, RsBridge, Streamable}
 import mongo4s.bson.direct.{DocumentCodecBridge, DriverCodecBridge, WireCodec}
-import mongo4s.bson.{BsonDecoder, BsonDocumentCodec, DecodeResult, FieldNaming}
+import mongo4s.bson.{BsonDecoder, BsonDocumentCodec, DecodeResult, FieldNaming, BsonDocumentDecoder}
 
 import scala.jdk.CollectionConverters.given
 
@@ -240,11 +240,11 @@ private[mongo4s] final class DirectMongoCollectionImpl[F[*], S[*], A](
     )
   end aggregateDirect
 
-  def aggregate[B](pipeline: Seq[Stage[A]])(using session: Option[ClientSession])(using codec: BsonDocumentCodec[B]): AggregateQuery[F, S, B] =
+  def aggregate[B](pipeline: Seq[Stage[A]])(using session: Option[ClientSession])(using decoder: BsonDocumentDecoder[B]): AggregateQuery[F, S, B] =
     AggregateQueryImpl(
       collection = underlying,
       pipeline = pipeline.map(_.toBson(naming)),
-      codec = codec,
+      decoder = decoder,
       allowDiskUse = None,
       session = session,
     )